# Conflicts:
#	.idea/misc.xml
#	.idea/webContexts.xml
#	src/main/resources/db.properties
#	src/main/webapp/common/admin/header.jsp
#	src/main/webapp/views/web/table.jsp
#	target/happyhome.war
#	target/happyhome/common/admin/header.jsp
#	target/happyhome/views/web/table.jsp

+++ src/main/java/controller/admin/CategoryControllerAdmin.java
@@ -1,7 +1,11 @@
 package controller.admin;
 
+import dao.RoleDAO;
 import model.CategoryProModel;
+import model.Role;
+import model.UserModel;
 import service.CategoryProService;
+import util.MessageUtil;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -12,16 +16,31 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(name = "CategoryController", value = "/data-category")
+@WebServlet(name = "CategoryController", value = "/admin-data-category")
 public class CategoryControllerAdmin extends HttpServlet {
+
+    String name = "List-Category-Product";
+    private static String editAccess = "sửa danh mục sản phẩm";
+    private static String deleteAccess = "xóa danh mục sản phẩm";
+    private static String addAccess = "thêm danh mục sản phẩm";
+    private static String listAccess = "xem danh sách danh mục sản phẩm";
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String action = "";
         if(request.getParameter("action") != null){
             action = request.getParameter("action");
         }
+        UserModel user = (UserModel) request.getSession().getAttribute("auth");
+        Role roleUser = RoleDAO.findById(user.getRole());
+        boolean access = access(action,roleUser);
+        boolean deletePm = Access.checkAccess(roleUser.getPermission(),RoleDAO.findIdPermissionByName(deleteAccess));
+        MessageUtil.showMessage(request);
         List<CategoryProModel> listCate;
         if(action.equals("add")){
+            if(!access){
+                response.sendRedirect("/admin-role-data?message=not_permission");
+                return;
+            }
             String name = request.getParameter("cate");
             CategoryProService.addCate(name);
             listCate = CategoryProService.findAll();
@@ -31,6 +50,10 @@ public class CategoryControllerAdmin extends HttpServlet {
             return;
         }
         if(action.equals("delete")){
+            if(!access){
+                response.sendRedirect("/admin-role-data?message=not_permission");
+                return;
+            }
             int id = Integer.parseInt(request.getParameter("id"));
             CategoryProService.deleteCate(id);
             listCate = CategoryProService.findAll();
@@ -40,6 +63,10 @@ public class CategoryControllerAdmin extends HttpServlet {
             return;
         }
         if(action.equals("edit")){
+            if(!access){
+                response.sendRedirect("/admin-role-data?message=not_permission");
+                return;
+            }
             String name = request.getParameter("cate");
             int id = Integer.parseInt(request.getParameter("id"));
             if(name == null){
@@ -55,8 +82,13 @@ public class CategoryControllerAdmin extends HttpServlet {
             request.getRequestDispatcher("views/admin/edit-category-product.jsp").forward(request,response);
             return;
         }
+        if(!access){
+            request.getRequestDispatcher("views/admin/no-permission.jsp").forward(request, response);
+            return;
+        }
         listCate = CategoryProService.findAll();
         request.setAttribute("listCate",listCate);
+        request.setAttribute("deletePm",deletePm);
         RequestDispatcher rd = request.getRequestDispatcher("views/admin/category-product.jsp");
         rd.forward(request,response);
     }
@@ -64,6 +96,19 @@ public class CategoryControllerAdmin extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
+    }
 
+    private static boolean access(String action, Role role) throws ServletException, IOException {
+        int access;
+        if(action.equals("edit")){
+            access = RoleDAO.findIdPermissionByName(editAccess);
+        } else if (action.equals("add")) {
+            access = RoleDAO.findIdPermissionByName(addAccess);
+        } else if (action.equals("delete")) {
+            access = RoleDAO.findIdPermissionByName(deleteAccess);
+        } else {
+            access = RoleDAO.findIdPermissionByName(listAccess);
+        }
+        return Access.checkAccess(role.getPermission(),access);
     }
 }
